parser: Lengths in TestGroupData Must be in Bytes
Currently, ivlen and returned_bits_len in the TestGroupData struct
were parsed without consideration of the fact that they are
represented in bits. This lead to downstream bugs where extra-large
IVs were being passed to AES-GCM with internal IVs.

This change fixes this issue by representing the lengths in bytes
in the parsed TestGroupData object.